+++ crates/wdl-engine/src/backend/docker.rs
@@ -474,7 +474,7 @@ impl TaskExecutionBackend for DockerBackend {
     #[cfg(unix)]
     fn cleanup<'a>(
         &'a self,
-        work_dir: &'a std::path::Path,
+        work_dir: &'a EvaluationPath,
         token: CancellationToken,
     ) -> Option<futures::future::BoxFuture<'a, ()>> {
         use futures::FutureExt;
@@ -487,6 +487,8 @@ impl TaskExecutionBackend for DockerBackend {
         /// Amount of memory to reserve for the cleanup task.
         const CLEANUP_MEMORY: f64 = 0.05;
 
+        // SAFETY: the work directory is always local for the Docker backend
+        let work_dir = work_dir.as_local().expect("path should be local");
         assert!(work_dir.is_absolute(), "work directory should be absolute");
 
         let backend = self.inner.clone();